version-1.1

+++ app/Http/Controllers/Backend/Trash/ModuleTrashController.php
@@ -6,11 +6,15 @@ use App\Models\Module;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Brian2694\Toastr\Facades\Toastr;
+use Illuminate\Support\Facades\Gate;
 
 class ModuleTrashController extends Controller
 {
     public function trash()
     {
+        //authorize this user to access/give access to admin dashboard
+        Gate::authorize('delete-module');
+
         $modules = Module::onlyTrashed()->latest('id')
         ->select(['id', 'module_name', 'module_slug', 'updated_at'])
         ->paginate(20);
@@ -20,6 +24,9 @@ class ModuleTrashController extends Controller
 
     public function restore($module_slug)
     {
+        //authorize this user to access/give access to admin dashboard
+        Gate::authorize('delete-module');
+
         $module = Module::onlyTrashed()->where('module_slug', $module_slug)->first();
         $module->restore();
 
@@ -29,6 +36,9 @@ class ModuleTrashController extends Controller
 
     public function forceDelete($module_slug)
     {
+        //authorize this user to access/give access to admin dashboard
+        Gate::authorize('delete-module');
+
         $module = Module::onlyTrashed()->where('module_slug', $module_slug)->first();
         $module->forceDelete();
